correct data sizes

+++ cube/common/address_resolution.c
@@ -1,6 +1,6 @@
 #include "address_resolution.h"
 
-int resolve_data_link_addr(int network_addr) {
+uint32_t resolve_data_link_addr(byte network_addr) {
     // We have, like, four devices total. Let's not make this hard on ourselves.
     int data_link_addr;
     switch (network_addr) {
@@ -20,9 +20,8 @@ int resolve_data_link_addr(int network_addr) {
     return data_link_addr;
 }
 
-int resolve_network_addr(int port) {
-    // They aren't so different after all!
+byte resolve_network_addr(byte port) {
     // This function exists in case you want to make the transport layer
-    // more elaborate.
+    // more elaborate. But for now...
     return port;
 }
