+++ cube/common/address_resolution.h
@@ -1,6 +1,8 @@
 #ifndef _ADDRESS_RESOLUTION_H
 #define _ADDRESS_RESOLUTION_H
 
+#include "networking_constants.h"
+#include <stdint.h>
 /*
     In a protocol stack, you will have a different kind of address for every
     layer. This is true for our protocol stack: our data link address is 4
@@ -11,7 +13,7 @@
     see how the TCP/IP stack does it. It's a little bizarre.
 */
 
-int resolve_data_link_addr(int network_addr);
-int resolve_network_addr(int port);
+uint32_t resolve_data_link_addr(byte network_addr);
+byte resolve_network_addr(byte port);
 
 #endif
